// ccm_react/src/client/models/GammaDistributionSpec.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A way to specify a gamma distribution by giving a shape and scale.
 */
export type GammaDistributionSpec = {
  type: 'gamma';
  distribution: 'gamma';
  clip?: any[];
  shape: number;
  scale: number;
};